+++ invenio_communities/communities/schema.py
@@ -51,7 +51,7 @@ def _not_blank(**kwargs):
             "Field cannot be blank or longer than {max_} characters.".format(max_=max_)
         ),
         min=1,
-        **kwargs
+        **kwargs,
     )
 
 
@@ -60,7 +60,7 @@ def no_longer_than(max, **kwargs):
     return validate.Length(
         error=_("Field cannot be longer than {max} characters.".format(max=max)),
         max=max,
-        **kwargs
+        **kwargs,
     )
 
 
@@ -163,15 +163,29 @@ class DeletionStatusSchema(Schema):
     status = fields.String(dump_only=True)
 
 
+class CommunityThemeStyleSchema(Schema):
+    """Community Theme configuration schema."""
+
+    font = fields.Dict()
+    primaryColor = fields.Str()
+    secondaryColor = fields.Str()
+    tertiaryColor = fields.Str()
+    primaryTextColor = fields.Str()
+    secondaryTextColor = fields.Str()
+    tertiaryTextColor = fields.Str()
+    mainHeaderBackgroundColor = fields.Str()
+
+
 class CommunityThemeSchema(Schema):
     """Community theme schema."""
 
-    config = fields.Dict()
+    style = fields.Nested(CommunityThemeStyleSchema)
     brand = fields.Str()
+    enabled = fields.Boolean()
 
 
-class CommunitySchema(BaseRecordSchema, FieldPermissionsMixin):
-    """Schema for the community metadata."""
+class BaseCommunitySchema(BaseRecordSchema, FieldPermissionsMixin):
+    """Base schema for the community metadata."""
 
     class Meta:
         """Meta attributes for the schema."""
@@ -243,6 +257,24 @@ class CommunitySchema(BaseRecordSchema, FieldPermissionsMixin):
         """Ensure slug is lowercase."""
         in_data["slug"] = in_data["slug"].lower()
         return in_data
+
+
+class CommunityParentSchema(BaseCommunitySchema):
+    """Community parent schema."""
+
+
+class CommunitySchema(BaseCommunitySchema):
+    """Community schema."""
+
+    parent = NestedAttribute(CommunityParentSchema, dump_only=True)
+
+    @post_dump
+    def post_dump(self, data, many, **kwargs):
+        """Hide parent field if it's not present."""
+        data = super().post_dump(data, many, **kwargs)
+        if data.get("parent") is None:
+            data.pop("parent", None)
+        return data
 
 
 class CommunityFeaturedSchema(Schema):