+++ invenio_communities/communities/services/results.py
@@ -112,7 +112,7 @@ class CommunityItem(RecordItem):
         return self._links_tpl.expand(self._identity, self._record)
 
     def to_dict(self):
-        """Get a dictionary for the request."""
+        """Get a dictionary for the community."""
         res = self.data
         if self._errors:
             res["errors"] = self._errors